Removed commented out codes. Changed draw_image() function to fix image corruption when the image is partially out of the screen. For-loop optimizations in display driver. Delta time value limit is increased from 8-bit to 32-bit, also no longer kernel panics when delta time is larger than 255. Renamed TYPE_NONE and RPN_TYPE_NONE to TYPE_NULL and RPN_TYPE_NULL respectively. Added key input as variable type to have easier access when using with instructions that accesses to the memory. Added ADDR_ARG_PTR for function arguments that are pointers, renamed ADDR_ARG to ADDR_ARG_ABS. Renamed MEM_BUF to VAR_BUFFER. Changed how function calls and returns work. Argument count no longer saved to MSBs of return address. Each return instruction now takes an operand to know how many arguments are needed to be popped from stack. Return value will be saved to a variable named func_return_val and will be assigned to destination after returned from function. Added get_key_held_for_time() function to handle long button presses in user friendly way (currently broken). Updated test code to test newly changed features.

+++ STM32F411CEU_Handheld_Console/Core/STM32_Handheld_Console/game_engine/game_engine.h
@@ -67,7 +67,7 @@ typedef struct s_ENGINE_SETTINGS
 extern ENGINE_SETTINGS engine_settings;
 
 extern uint32_t elapsed_time;
-extern uint8_t delta_time;
+extern uint32_t delta_time;
 
 void wait_frame_time(void);
 void load_game(void *game);
